requirements.txt
streamlit
tensorflow
numpy
Pillow
opencv-python
scikit-learn
pandas
matplotlib
seaborn
nltk
